Move ProductCard.js into components folder and delete ProductCard.module.css as it is no longer needed
