--- src/components/ProductCard/ProductCard.module.css
@@ -1,67 +1,0 @@
-.image {
-  width: 200px;
-  height: 300px;
-  object-fit: cover;
-  border-radius: 20px;
-  margin: 0px;
-}
-
-.wrapper {
-  display: grid;
-  justify-content: center;
-  align-items: center;
-  width: 200px;
-  border-radius: 20px;
-  gap: 10px;
-  cursor: pointer;
-  position: relative;
-  box-shadow: 0px 20px 40px rgba(52, 53, 99, 0.2),
-    0px 1px 3px rgba(0, 0, 0, 0.5);
-}
-
-.text-wrapper {
-  position: absolute;
-  bottom: 0px;
-  left: 0px;
-  border-radius: 0 0 20px 20px;
-  background: rgba(255, 255, 255, 0.2);
-  width: 200px;
-  padding: 10px 0;
-  backdrop-filter: blur(40px);
-}
-
-.title {
-  font-weight: 600;
-  text-align: center;
-  margin: 0;
-  color: #014c40;
-}
-
-.price {
-  font-weight: normal;
-  text-align: center;
-  margin: 0;
-}
-
-.add-button {
-  position: absolute;
-  top: 20px;
-  right: 20px;
-  background-color: #014c40;
-  padding: 10px;
-  width: 40px;
-  height: 40px;
-  border-radius: 50%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  color: #fff;
-  font-weight: bold;
-  margin: 0;
-  border: 3px solid white;
-}
-
-.add-button:hover {
-  transform: scale(1.2);
-  transition: 0.2s;
-}
